// apps/web/generated/api/models/freelancerApprovalStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */

export type FreelancerApprovalStatus = typeof FreelancerApprovalStatus[keyof typeof FreelancerApprovalStatus];


export const FreelancerApprovalStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  suspended: 'suspended',
} as const;